if the file exists, don't download it

+++ cos/packages/archive.lua
@@ -7,16 +7,17 @@ archive.metadata = {
     dependencies = {"libdeflate"}
 }
 archive.startup = function()
-    local hweb, err = http.get("https://raw.githubusercontent.com/MCJack123/CC-Archive/master/archive.lua")
-    if not hweb then
-        log("Failed to download archive.lua: " .. err, false, "error")
-        error("Failed to download archive.lua: " .. err)
+    if not fs.exists("/cos/lib/archive.lua") then
+        local hweb, err = http.get("https://raw.githubusercontent.com/MCJack123/CC-Archive/master/archive.lua")
+        if not hweb then
+            log("Failed to download archive.lua: " .. err, false, "error")
+            error("Failed to download archive.lua: " .. err)
+        end
+        local hfile = fs.open("/cos/lib/archive.lua", "w")
+        hfile.write(hweb.readAll())
+        hfile.close()
+        hweb.close()
     end
-    local hfile = fs.open("/cos/lib/archive.lua", "w")
-    hfile.write(hweb.readAll())
-    hfile.close()
-    hweb.close()
-
     return true
 end
 return archive
